[css-pseudo] Move all ::marker style adjustments into StyleAdjuster
LayoutNGListItem was applying some style adjustments to markers, like
'white-space: pre' to preserve trailing spaces in outside markers.
But this was only affecting markers with 'content: normal', and the
changes weren't exposed in getComputedStyle().
This patch moves these adjustments into StyleAdjuster, so that all
markers are affected equally.

BUG=457718

TEST=external/wpt/css/css-pseudo/marker-content-018.html

The test fails in legacy because the 'content' property is not supported
yet in ::marker.
There is a new failure in marker-supported-properties.html because now
'white-space' will compute to 'pre' in outside markers, even if you
specify a different value. But normal markers were already enforcing
'pre' at used value time anyways.

Change-Id: I6d8b8ec7fb26cb04402032924f87b2c3b8fdb9cd